Adafruit Wire library already defines TinyWireM as Wire
So, this was a little tricky.  For Adafruit Trinket and Gemma, you want
to remove the definition of Wire as TinyWireM.  If you don’t other uses
of Wire in a sketch get messed up.
But, the Arduino Gemma doesn’t use the Adafruit Wire library.  So
Arduino Gemma compiles would fail.  I made ARDUINO_AVR_GEMMA a special
case.
Version 1.0.3