Delete examples/keypad/doc directory
Did not intend for the documentation to be pulled into the pull request. Typically Arduino library examples do not contain separate documentation. All documentation is in the code file not separated out like I had it here. If someone would like the Fritizing file or other documentation, feel free to contact me.